test: move test-support.ts up one level
